+++ src/main/java/com/techelevator/models/Drink.java
@@ -2,9 +2,11 @@ package com.techelevator.models;
 
 import com.techelevator.interfaces.TypeConstants;
 
+import java.math.BigDecimal;
+
 public class Drink extends Product implements TypeConstants {
 
-    public Drink(String name, double price) {
+    public Drink(String name, BigDecimal price) {
         super(name, price, TYPE_DRINK);
     }
