+++ src/main/java/com/techelevator/models/Gum.java
@@ -2,9 +2,11 @@ package com.techelevator.models;
 
 import com.techelevator.interfaces.TypeConstants;
 
+import java.math.BigDecimal;
+
 public class Gum extends Product implements TypeConstants {
 
-    public Gum(String name, double price) {
+    public Gum(String name, BigDecimal price) {
         super(name, price, TYPE_GUM);
     }
